f2000 moved NavBar to folder, added SearchBar to same folder
